Start converting registry views to use openraft
