--- src.old/views/mod.rs
@@ -1,5 +1,0 @@
-pub(crate) mod blobs;
-pub(crate) mod get;
-pub(crate) mod manifests;
-pub(crate) mod tags;
-pub(crate) mod utils;
