+++ src/registry/blobs/mod.rs
