+++ src/registry/blobs/uploads/get.rs
@@ -1,28 +1,20 @@
 use crate::config::Configuration;
-use crate::headers::ContentRange;
 use crate::headers::Token;
-
 use crate::types::RepositoryName;
 use crate::utils::get_upload_path;
-use rocket::data::Data;
 use rocket::http::Header;
 use rocket::http::Status;
 use rocket::request::Request;
 use rocket::response::{Responder, Response};
 use rocket::State;
 use std::io::Cursor;
-
 pub(crate) enum Responses {
     MustAuthenticate {
         challenge: String,
     },
     AccessDenied {},
     UploadInvalid {},
-    RangeNotSatisfiable {
-        repository: RepositoryName,
-        upload_id: String,
-        size: u64,
-    },
+    UploadNotFound {},
     Ok {
         repository: RepositoryName,
         upload_id: String,
@@ -34,7 +26,7 @@ impl<'r> Responder<'r, 'static> for Responses {
     fn respond_to(self, _req: &Request) -> Result<Response<'static>, Status> {
         match self {
             Responses::MustAuthenticate { challenge } => {
-                let body = crate::views::utils::simple_oci_error(
+                let body = crate::registry::utils::simple_oci_error(
                     "UNAUTHORIZED",
                     "authentication required",
                 );
@@ -46,7 +38,7 @@ impl<'r> Responder<'r, 'static> for Responses {
                     .ok()
             }
             Responses::AccessDenied {} => {
-                let body = crate::views::utils::simple_oci_error(
+                let body = crate::registry::utils::simple_oci_error(
                     "DENIED",
                     "requested access to the resource is denied",
                 );
@@ -57,7 +49,7 @@ impl<'r> Responder<'r, 'static> for Responses {
                     .ok()
             }
             Responses::UploadInvalid {} => {
-                let body = crate::views::utils::simple_oci_error(
+                let body = crate::registry::utils::simple_oci_error(
                     "BLOB_UPLOAD_INVALID",
                     "the upload was invalid",
                 );
@@ -67,31 +59,15 @@ impl<'r> Responder<'r, 'static> for Responses {
                     .status(Status::BadRequest)
                     .ok()
             }
-            Responses::RangeNotSatisfiable {
-                repository,
-                upload_id,
-                size,
-            } => {
-                /*
-                416 Range Not Satisfiable
-                Location: /v2/<name>/blobs/uploads/<uuid>
-                Range: 0-<offset>
-                Content-Length: 0
-                Docker-Upload-UUID: <uuid>
-                */
-
-                let range_end = if size > 0 { size - 1 } else { 0 };
-
+            Responses::UploadNotFound {} => {
+                let body = crate::registry::utils::simple_oci_error(
+                    "BLOB_UPLOAD_UNKNOWN",
+                    "blob upload unknown to registry",
+                );
                 Response::build()
-                    .header(Header::new(
-                        "Location",
-                        format!("/v2/{repository}/blobs/uploads/{upload_id}"),
-                    ))
-                    .header(Header::new("Range", format!("0-{range_end}")))
-                    .header(Header::new("Content-Length", "0"))
-                    .header(Header::new("Blob-Upload-Session-ID", upload_id.clone()))
-                    .header(Header::new("Docker-Upload-UUID", upload_id))
-                    .status(Status::RangeNotSatisfiable)
+                    .header(Header::new("Content-Length", body.len().to_string()))
+                    .sized_body(body.len(), Cursor::new(body))
+                    .status(Status::NotFound)
                     .ok()
             }
             Responses::Ok {
@@ -125,14 +101,12 @@ impl<'r> Responder<'r, 'static> for Responses {
     }
 }
 
-#[patch("/<repository>/blobs/uploads/<upload_id>", data = "<body>")]
-pub(crate) async fn patch(
+#[get("/<repository>/blobs/uploads/<upload_id>")]
+pub(crate) async fn get(
     repository: RepositoryName,
     upload_id: String,
-    range: Option<ContentRange>,
     config: &State<Configuration>,
     token: Token,
-    body: Data<'_>,
 ) -> Responses {
     let config: &Configuration = config.inner();
 
@@ -142,41 +116,13 @@ pub(crate) async fn patch(
         };
     }
 
-    if !token.has_permission(&repository, "push") {
+    if !token.has_permission(&repository, "pull") {
         return Responses::AccessDenied {};
     }
 
     let filename = get_upload_path(&config.storage, &upload_id);
-
     if !filename.is_file() {
-        return Responses::UploadInvalid {};
-    }
-
-    if let Some(range) = range {
-        let size = match tokio::fs::metadata(&filename).await {
-            Ok(value) => value.len(),
-            _ => 0,
-        };
-
-        if range.stop < range.start {
-            return Responses::RangeNotSatisfiable {
-                repository,
-                upload_id,
-                size,
-            };
-        }
-
-        if range.start != size {
-            return Responses::RangeNotSatisfiable {
-                repository,
-                upload_id,
-                size,
-            };
-        }
-    }
-
-    if !crate::views::utils::upload_part(&filename, body).await {
-        return Responses::UploadInvalid {};
+        return Responses::UploadNotFound {};
     }
 
     let size = match tokio::fs::metadata(filename).await {
