+++ src/registry/blobs/uploads/patch.rs
@@ -1,20 +1,28 @@
 use crate::config::Configuration;
+use crate::headers::ContentRange;
 use crate::headers::Token;
+
 use crate::types::RepositoryName;
 use crate::utils::get_upload_path;
+use rocket::data::Data;
 use rocket::http::Header;
 use rocket::http::Status;
 use rocket::request::Request;
 use rocket::response::{Responder, Response};
 use rocket::State;
 use std::io::Cursor;
+
 pub(crate) enum Responses {
     MustAuthenticate {
         challenge: String,
     },
     AccessDenied {},
     UploadInvalid {},
-    UploadNotFound {},
+    RangeNotSatisfiable {
+        repository: RepositoryName,
+        upload_id: String,
+        size: u64,
+    },
     Ok {
         repository: RepositoryName,
         upload_id: String,
@@ -26,7 +34,7 @@ impl<'r> Responder<'r, 'static> for Responses {
     fn respond_to(self, _req: &Request) -> Result<Response<'static>, Status> {
         match self {
             Responses::MustAuthenticate { challenge } => {
-                let body = crate::views::utils::simple_oci_error(
+                let body = crate::registry::utils::simple_oci_error(
                     "UNAUTHORIZED",
                     "authentication required",
                 );
@@ -38,7 +46,7 @@ impl<'r> Responder<'r, 'static> for Responses {
                     .ok()
             }
             Responses::AccessDenied {} => {
-                let body = crate::views::utils::simple_oci_error(
+                let body = crate::registry::utils::simple_oci_error(
                     "DENIED",
                     "requested access to the resource is denied",
                 );
@@ -49,7 +57,7 @@ impl<'r> Responder<'r, 'static> for Responses {
                     .ok()
             }
             Responses::UploadInvalid {} => {
-                let body = crate::views::utils::simple_oci_error(
+                let body = crate::registry::utils::simple_oci_error(
                     "BLOB_UPLOAD_INVALID",
                     "the upload was invalid",
                 );
@@ -59,15 +67,31 @@ impl<'r> Responder<'r, 'static> for Responses {
                     .status(Status::BadRequest)
                     .ok()
             }
-            Responses::UploadNotFound {} => {
-                let body = crate::views::utils::simple_oci_error(
-                    "BLOB_UPLOAD_UNKNOWN",
-                    "blob upload unknown to registry",
-                );
+            Responses::RangeNotSatisfiable {
+                repository,
+                upload_id,
+                size,
+            } => {
+                /*
+                416 Range Not Satisfiable
+                Location: /v2/<name>/blobs/uploads/<uuid>
+                Range: 0-<offset>
+                Content-Length: 0
+                Docker-Upload-UUID: <uuid>
+                */
+
+                let range_end = if size > 0 { size - 1 } else { 0 };
+
                 Response::build()
-                    .header(Header::new("Content-Length", body.len().to_string()))
-                    .sized_body(body.len(), Cursor::new(body))
-                    .status(Status::NotFound)
+                    .header(Header::new(
+                        "Location",
+                        format!("/v2/{repository}/blobs/uploads/{upload_id}"),
+                    ))
+                    .header(Header::new("Range", format!("0-{range_end}")))
+                    .header(Header::new("Content-Length", "0"))
+                    .header(Header::new("Blob-Upload-Session-ID", upload_id.clone()))
+                    .header(Header::new("Docker-Upload-UUID", upload_id))
+                    .status(Status::RangeNotSatisfiable)
                     .ok()
             }
             Responses::Ok {
@@ -101,12 +125,14 @@ impl<'r> Responder<'r, 'static> for Responses {
     }
 }
 
-#[get("/<repository>/blobs/uploads/<upload_id>")]
-pub(crate) async fn get(
+#[patch("/<repository>/blobs/uploads/<upload_id>", data = "<body>")]
+pub(crate) async fn patch(
     repository: RepositoryName,
     upload_id: String,
+    range: Option<ContentRange>,
     config: &State<Configuration>,
     token: Token,
+    body: Data<'_>,
 ) -> Responses {
     let config: &Configuration = config.inner();
 
@@ -116,13 +142,41 @@ pub(crate) async fn get(
         };
     }
 
-    if !token.has_permission(&repository, "pull") {
+    if !token.has_permission(&repository, "push") {
         return Responses::AccessDenied {};
     }
 
     let filename = get_upload_path(&config.storage, &upload_id);
+
     if !filename.is_file() {
-        return Responses::UploadNotFound {};
+        return Responses::UploadInvalid {};
+    }
+
+    if let Some(range) = range {
+        let size = match tokio::fs::metadata(&filename).await {
+            Ok(value) => value.len(),
+            _ => 0,
+        };
+
+        if range.stop < range.start {
+            return Responses::RangeNotSatisfiable {
+                repository,
+                upload_id,
+                size,
+            };
+        }
+
+        if range.start != size {
+            return Responses::RangeNotSatisfiable {
+                repository,
+                upload_id,
+                size,
+            };
+        }
+    }
+
+    if !crate::registry::utils::upload_part(&filename, body).await {
+        return Responses::UploadInvalid {};
     }
 
     let size = match tokio::fs::metadata(filename).await {
