+++ src/registry/blobs/uploads/put.rs
@@ -1,15 +1,12 @@
 use crate::config::Configuration;
-use crate::extractor::Extractor;
-use crate::headers::ContentType;
 use crate::headers::Token;
 use crate::rpc::RpcClient;
 use crate::types::Digest;
 use crate::types::RegistryAction;
 use crate::types::RegistryState;
 use crate::types::RepositoryName;
-use crate::utils::get_manifest_path;
-use crate::views::utils::get_hash;
-use crate::webhook::Event;
+use crate::utils::get_blob_path;
+use crate::utils::get_upload_path;
 use chrono::prelude::*;
 use rocket::data::Data;
 use rocket::http::Header;
@@ -25,6 +22,7 @@ pub(crate) enum Responses {
         challenge: String,
     },
     AccessDenied {},
+    DigestInvalid {},
     UploadInvalid {},
     Ok {
         repository: RepositoryName,
@@ -36,7 +34,7 @@ impl<'r> Responder<'r, 'static> for Responses {
     fn respond_to(self, _req: &Request) -> Result<Response<'static>, Status> {
         match self {
             Responses::MustAuthenticate { challenge } => {
-                let body = crate::views::utils::simple_oci_error(
+                let body = crate::registry::utils::simple_oci_error(
                     "UNAUTHORIZED",
                     "authentication required",
                 );
@@ -48,7 +46,7 @@ impl<'r> Responder<'r, 'static> for Responses {
                     .ok()
             }
             Responses::AccessDenied {} => {
-                let body = crate::views::utils::simple_oci_error(
+                let body = crate::registry::utils::simple_oci_error(
                     "DENIED",
                     "requested access to the resource is denied",
                 );
@@ -58,9 +56,22 @@ impl<'r> Responder<'r, 'static> for Responses {
                     .status(Status::Forbidden)
                     .ok()
             }
+            Responses::DigestInvalid {} => {
+                let body = crate::registry::utils::simple_oci_error(
+                    "DIGEST_INVALID",
+                    "provided digest did not match uploaded content",
+                );
+                Response::build()
+                    .header(Header::new("Content-Length", body.len().to_string()))
+                    .sized_body(body.len(), Cursor::new(body))
+                    .status(Status::BadRequest)
+                    .ok()
+            }
             Responses::UploadInvalid {} => {
-                let body =
-                    crate::views::utils::simple_oci_error("MANIFEST_INVALID", "upload was invalid");
+                let body = crate::registry::utils::simple_oci_error(
+                    "BLOB_UPLOAD_INVALID",
+                    "the upload was invalid",
+                );
                 Response::build()
                     .header(Header::new("Content-Length", body.len().to_string()))
                     .sized_body(body.len(), Cursor::new(body))
@@ -70,21 +81,20 @@ impl<'r> Responder<'r, 'static> for Responses {
             Responses::Ok { repository, digest } => {
                 /*
                 201 Created
-                Location: <url>
+                Location: <blob location>
+                Content-Range: <start of range>-<end of range, inclusive>
                 Content-Length: 0
                 Docker-Content-Digest: <digest>
                 */
-                let location = Header::new(
-                    "Location",
-                    format!("/v2/{}/manifests/{}", repository, digest),
-                );
-                let length = Header::new("Content-Length", "0");
-                let upload_uuid = Header::new("Docker-Content-Digest", digest.to_string());
 
                 Response::build()
-                    .header(location)
-                    .header(length)
-                    .header(upload_uuid)
+                    .header(Header::new(
+                        "Location",
+                        format!("/v2/{repository}/blobs/{digest}"),
+                    ))
+                    .header(Header::new("Range", "0-0"))
+                    .header(Header::new("Content-Length", "0"))
+                    .header(Header::new("Docker-Content-Digest", digest.to_string()))
                     .status(Status::Created)
                     .ok()
             }
@@ -92,21 +102,19 @@ impl<'r> Responder<'r, 'static> for Responses {
     }
 }
 
-#[put("/<repository>/manifests/<tag>", data = "<body>")]
+#[put("/<repository>/blobs/uploads/<upload_id>?<digest>", data = "<body>")]
 pub(crate) async fn put(
     repository: RepositoryName,
-    tag: String,
+    upload_id: String,
+    digest: Digest,
     config: &State<Configuration>,
     state: &State<Arc<RegistryState>>,
     submission: &State<Arc<RpcClient>>,
-    extractor: &State<Extractor>,
-    content_type: ContentType,
     token: Token,
     body: Data<'_>,
 ) -> Responses {
     let config: &Configuration = config.inner();
     let state: &RegistryState = state.inner();
-    let extractor: &Extractor = extractor.inner();
     let submission: &RpcClient = submission.inner();
 
     if !token.validated_token {
@@ -119,92 +127,64 @@ pub(crate) async fn put(
         return Responses::AccessDenied {};
     }
 
-    let upload_path = crate::utils::get_temp_path(&config.storage);
-
-    if !crate::views::utils::upload_part(&upload_path, body).await {
+    if digest.algo != "sha256" {
         return Responses::UploadInvalid {};
     }
 
-    let size = match tokio::fs::metadata(&upload_path).await {
-        Ok(result) => result.len(),
+    let filename = get_upload_path(&config.storage, &upload_id);
+
+    if !filename.is_file() {
+        return Responses::UploadInvalid {};
+    }
+
+    if !crate::registry::utils::upload_part(&filename, body).await {
+        return Responses::UploadInvalid {};
+    }
+
+    // Validate upload
+    if !crate::registry::utils::validate_hash(&filename, &digest).await {
+        return Responses::DigestInvalid {};
+    }
+
+    let dest = get_blob_path(&config.storage, &digest);
+
+    let stat = match tokio::fs::metadata(&filename).await {
+        Ok(result) => result,
         Err(_) => {
             return Responses::UploadInvalid {};
         }
     };
 
-    let digest = match get_hash(&upload_path).await {
-        Some(digest) => digest,
-        _ => {
+    match std::fs::rename(filename.clone(), dest.clone()) {
+        Ok(_) => {}
+        Err(_e) => {
             return Responses::UploadInvalid {};
         }
-    };
+    }
 
-    let extracted = extractor
-        .extract(
-            state,
-            &repository,
-            &digest,
-            &content_type.content_type,
-            &upload_path,
-        )
-        .await;
-
-    let mut actions = vec![
-        RegistryAction::ManifestMounted {
+    let actions = vec![
+        RegistryAction::BlobMounted {
             timestamp: Utc::now(),
             digest: digest.clone(),
             repository: repository.clone(),
             user: token.sub.clone(),
         },
-        RegistryAction::ManifestStored {
+        RegistryAction::BlobStat {
+            timestamp: Utc::now(),
+            digest: digest.clone(),
+            size: stat.len(),
+        },
+        RegistryAction::BlobStored {
             timestamp: Utc::now(),
             digest: digest.clone(),
             location: state.machine_identifier.clone(),
             user: token.sub.clone(),
         },
-        RegistryAction::ManifestStat {
-            timestamp: Utc::now(),
-            digest: digest.clone(),
-            size,
-        },
     ];
-
-    let extracted = match extracted {
-        Ok(extracted_actions) => extracted_actions,
-        _ => {
-            return Responses::UploadInvalid {};
-        }
-    };
-    actions.append(&mut extracted.clone());
-    actions.append(&mut vec![RegistryAction::HashTagged {
-        timestamp: Utc::now(),
-        repository: repository.clone(),
-        digest: digest.clone(),
-        tag: tag.clone(),
-        user: token.sub.clone(),
-    }]);
-
-    let dest = get_manifest_path(&config.storage, &digest);
-
-    match std::fs::rename(upload_path, dest) {
-        Ok(_) => {}
-        Err(_) => {
-            return Responses::UploadInvalid {};
-        }
-    }
 
     if !submission.send(actions).await {
         return Responses::UploadInvalid {};
     }
-
-    state
-        .send_webhook(Event {
-            repository: repository.clone(),
-            digest: digest.clone(),
-            tag,
-            content_type: content_type.content_type,
-        })
-        .await;
 
     Responses::Ok { repository, digest }
 }
