+++ src/registry/manifests/mod.rs
