+++ src/registry/manifests/put.rs
@@ -1,12 +1,15 @@
 use crate::config::Configuration;
+use crate::extractor::Extractor;
+use crate::headers::ContentType;
 use crate::headers::Token;
 use crate::rpc::RpcClient;
 use crate::types::Digest;
 use crate::types::RegistryAction;
 use crate::types::RegistryState;
 use crate::types::RepositoryName;
-use crate::utils::get_blob_path;
-use crate::utils::get_upload_path;
+use crate::utils::get_manifest_path;
+use crate::registry::utils::get_hash;
+use crate::webhook::Event;
 use chrono::prelude::*;
 use rocket::data::Data;
 use rocket::http::Header;
@@ -22,7 +25,6 @@ pub(crate) enum Responses {
         challenge: String,
     },
     AccessDenied {},
-    DigestInvalid {},
     UploadInvalid {},
     Ok {
         repository: RepositoryName,
@@ -34,7 +36,7 @@ impl<'r> Responder<'r, 'static> for Responses {
     fn respond_to(self, _req: &Request) -> Result<Response<'static>, Status> {
         match self {
             Responses::MustAuthenticate { challenge } => {
-                let body = crate::views::utils::simple_oci_error(
+                let body = crate::registry::utils::simple_oci_error(
                     "UNAUTHORIZED",
                     "authentication required",
                 );
@@ -46,7 +48,7 @@ impl<'r> Responder<'r, 'static> for Responses {
                     .ok()
             }
             Responses::AccessDenied {} => {
-                let body = crate::views::utils::simple_oci_error(
+                let body = crate::registry::utils::simple_oci_error(
                     "DENIED",
                     "requested access to the resource is denied",
                 );
@@ -56,22 +58,9 @@ impl<'r> Responder<'r, 'static> for Responses {
                     .status(Status::Forbidden)
                     .ok()
             }
-            Responses::DigestInvalid {} => {
-                let body = crate::views::utils::simple_oci_error(
-                    "DIGEST_INVALID",
-                    "provided digest did not match uploaded content",
-                );
-                Response::build()
-                    .header(Header::new("Content-Length", body.len().to_string()))
-                    .sized_body(body.len(), Cursor::new(body))
-                    .status(Status::BadRequest)
-                    .ok()
-            }
             Responses::UploadInvalid {} => {
-                let body = crate::views::utils::simple_oci_error(
-                    "BLOB_UPLOAD_INVALID",
-                    "the upload was invalid",
-                );
+                let body =
+                    crate::registry::utils::simple_oci_error("MANIFEST_INVALID", "upload was invalid");
                 Response::build()
                     .header(Header::new("Content-Length", body.len().to_string()))
                     .sized_body(body.len(), Cursor::new(body))
@@ -81,20 +70,21 @@ impl<'r> Responder<'r, 'static> for Responses {
             Responses::Ok { repository, digest } => {
                 /*
                 201 Created
-                Location: <blob location>
-                Content-Range: <start of range>-<end of range, inclusive>
+                Location: <url>
                 Content-Length: 0
                 Docker-Content-Digest: <digest>
                 */
+                let location = Header::new(
+                    "Location",
+                    format!("/v2/{}/manifests/{}", repository, digest),
+                );
+                let length = Header::new("Content-Length", "0");
+                let upload_uuid = Header::new("Docker-Content-Digest", digest.to_string());
 
                 Response::build()
-                    .header(Header::new(
-                        "Location",
-                        format!("/v2/{repository}/blobs/{digest}"),
-                    ))
-                    .header(Header::new("Range", "0-0"))
-                    .header(Header::new("Content-Length", "0"))
-                    .header(Header::new("Docker-Content-Digest", digest.to_string()))
+                    .header(location)
+                    .header(length)
+                    .header(upload_uuid)
                     .status(Status::Created)
                     .ok()
             }
@@ -102,19 +92,21 @@ impl<'r> Responder<'r, 'static> for Responses {
     }
 }
 
-#[put("/<repository>/blobs/uploads/<upload_id>?<digest>", data = "<body>")]
+#[put("/<repository>/manifests/<tag>", data = "<body>")]
 pub(crate) async fn put(
     repository: RepositoryName,
-    upload_id: String,
-    digest: Digest,
+    tag: String,
     config: &State<Configuration>,
     state: &State<Arc<RegistryState>>,
     submission: &State<Arc<RpcClient>>,
+    extractor: &State<Extractor>,
+    content_type: ContentType,
     token: Token,
     body: Data<'_>,
 ) -> Responses {
     let config: &Configuration = config.inner();
     let state: &RegistryState = state.inner();
+    let extractor: &Extractor = extractor.inner();
     let submission: &RpcClient = submission.inner();
 
     if !token.validated_token {
@@ -127,64 +119,92 @@ pub(crate) async fn put(
         return Responses::AccessDenied {};
     }
 
-    if digest.algo != "sha256" {
+    let upload_path = crate::utils::get_temp_path(&config.storage);
+
+    if !crate::registry::utils::upload_part(&upload_path, body).await {
         return Responses::UploadInvalid {};
     }
 
-    let filename = get_upload_path(&config.storage, &upload_id);
-
-    if !filename.is_file() {
-        return Responses::UploadInvalid {};
-    }
-
-    if !crate::views::utils::upload_part(&filename, body).await {
-        return Responses::UploadInvalid {};
-    }
-
-    // Validate upload
-    if !crate::views::utils::validate_hash(&filename, &digest).await {
-        return Responses::DigestInvalid {};
-    }
-
-    let dest = get_blob_path(&config.storage, &digest);
-
-    let stat = match tokio::fs::metadata(&filename).await {
-        Ok(result) => result,
+    let size = match tokio::fs::metadata(&upload_path).await {
+        Ok(result) => result.len(),
         Err(_) => {
             return Responses::UploadInvalid {};
         }
     };
 
-    match std::fs::rename(filename.clone(), dest.clone()) {
-        Ok(_) => {}
-        Err(_e) => {
+    let digest = match get_hash(&upload_path).await {
+        Some(digest) => digest,
+        _ => {
             return Responses::UploadInvalid {};
         }
-    }
+    };
 
-    let actions = vec![
-        RegistryAction::BlobMounted {
+    let extracted = extractor
+        .extract(
+            state,
+            &repository,
+            &digest,
+            &content_type.content_type,
+            &upload_path,
+        )
+        .await;
+
+    let mut actions = vec![
+        RegistryAction::ManifestMounted {
             timestamp: Utc::now(),
             digest: digest.clone(),
             repository: repository.clone(),
             user: token.sub.clone(),
         },
-        RegistryAction::BlobStat {
-            timestamp: Utc::now(),
-            digest: digest.clone(),
-            size: stat.len(),
-        },
-        RegistryAction::BlobStored {
+        RegistryAction::ManifestStored {
             timestamp: Utc::now(),
             digest: digest.clone(),
             location: state.machine_identifier.clone(),
             user: token.sub.clone(),
         },
+        RegistryAction::ManifestStat {
+            timestamp: Utc::now(),
+            digest: digest.clone(),
+            size,
+        },
     ];
+
+    let extracted = match extracted {
+        Ok(extracted_actions) => extracted_actions,
+        _ => {
+            return Responses::UploadInvalid {};
+        }
+    };
+    actions.append(&mut extracted.clone());
+    actions.append(&mut vec![RegistryAction::HashTagged {
+        timestamp: Utc::now(),
+        repository: repository.clone(),
+        digest: digest.clone(),
+        tag: tag.clone(),
+        user: token.sub.clone(),
+    }]);
+
+    let dest = get_manifest_path(&config.storage, &digest);
+
+    match std::fs::rename(upload_path, dest) {
+        Ok(_) => {}
+        Err(_) => {
+            return Responses::UploadInvalid {};
+        }
+    }
 
     if !submission.send(actions).await {
         return Responses::UploadInvalid {};
     }
+
+    state
+        .send_webhook(Event {
+            repository: repository.clone(),
+            digest: digest.clone(),
+            tag,
+            content_type: content_type.content_type,
+        })
+        .await;
 
     Responses::Ok { repository, digest }
 }
